+++ routes/web.php
@@ -1,6 +1,7 @@
 <?php
 
 use App\Http\Controllers\content\HomeController;
+use App\Http\Controllers\content\ListProductController;
 use Illuminate\Support\Facades\Route;
 
 /*
@@ -23,7 +24,13 @@ Auth::routes();
 
 
 
-
+//Home
 Route::get('home',[HomeController::class ,'home']);
+Route::get('home/product-category/{id}',[HomeController::class, 'productCategory'])->name('home-product-category');
 
 Route::get('test',[HomeController::class ,'test']);
+
+// Show Product 
+Route::get('list-product/category',[ListProductController::class, 'productCategories'])->name('list-product-category');
+Route::get('list-product/category/{cate}/brand/{brand}',[ListProductController::class, 'productBrands'])->name('list-product-brand');
+
